feat(email-2fa): deepen retrieval service and remove pass-through
